# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangeMemberRoleDto": {
        "properties": {
          "role": {
            "enum": ["TENANT_ADMIN", "EMPLOYEE"],
            "type": "string"
          }
        },
        "required": ["role"],
        "type": "object"
      },
      "ChangeMemberStatusDto": {
        "properties": {
          "status": {
            "enum": ["ACTIVE", "DISABLED"],
            "type": "string"
          }
        },
        "required": ["status"],
        "type": "object"
      },
...
      "CreateInvitationDto": {
        "properties": {
          "phone": {
            "example": "13800138000",
            "type": "string"
          },
          "role": {
            "enum": ["TENANT_ADMIN", "EMPLOYEE"],
            "type": "string"
          }
        },
        "required": ["phone", "role"],
        "type": "object"
      },
...
      "CreatedInvitationResponseDto": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": ["PENDING"],
            "type": "string"
          }
        },
        "required": ["id", "status"],
        "type": "object"
      },
...
      "InvitationPageResponseDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TenantInvitationResponseDto"
            },
            "type": "array"
          },
          "nextCursor": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": ["items"],
        "type": "object"
      },
      "InvitationResentResponseDto": {
        "properties": {
          "accepted": {
            "example": true,
            "type": "boolean"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": ["accepted", "expiresAt"],
        "type": "object"
      },
...
      "MemberObjectAccessDto": {
        "properties": {
          "canCreate": {
            "type": "boolean"
          },
          "canRead": {
            "type": "boolean"
          },
          "canUpdate": {
            "type": "boolean"
          },
          "mode": {
            "enum": ["INHERIT", "OVERRIDE"],
            "type": "string"
          },
          "readScope": {
            "enum": ["ALL", "OWN", "NONE"],
            "type": "string"
          },
          "updateScope": {
            "enum": ["ALL", "OWN", "NONE"],
            "type": "string"
          }
        },
        "required": ["mode"],
        "type": "object"
      },
      "MemberObjectAccessResponseDto": {
        "properties": {
          "effective": {
            "$ref": "#/components/schemas/MemberObjectPolicyResponseDto"
          },
          "inherited": {
            "$ref": "#/components/schemas/MemberObjectPolicyResponseDto"
          },
          "mode": {
            "enum": ["INHERIT", "OVERRIDE"],
            "type": "string"
          },
          "objectCode": {
            "type": "string"
          },
          "objectId": {
            "format": "uuid",
            "type": "string"
          },
          "objectName": {
            "type": "string"
          },
          "override": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MemberObjectPolicyResponseDto"
              }
            ],
            "nullable": true,
            "type": "object"
          }
        },
        "required": [
          "objectId",
          "objectCode",
          "objectName",
          "mode",
          "inherited",
          "override",
          "effective"
        ],
        "type": "object"
      },
      "MemberObjectPolicyResponseDto": {
        "properties": {
          "canCreate": {
            "type": "boolean"
          },
          "canDelete": {
            "enum": [false],
            "type": "boolean"
          },
          "canRead": {
            "type": "boolean"
          },
          "canUpdate": {
            "type": "boolean"
          },
          "readScope": {
            "enum": ["ALL", "OWN", "NONE"],
            "type": "string"
          },
          "updateScope": {
            "enum": ["ALL", "OWN", "NONE"],
            "type": "string"
          }
        },
        "required": [
          "canCreate",
          "canRead",
          "canUpdate",
          "canDelete",
          "readScope",
          "updateScope"
        ],
        "type": "object"
      },
      "MemberRecipientDto": {
        "properties": {
          "recipientMemberId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": ["recipientMemberId"],
        "type": "object"
      },
      "MembershipActionResponseDto": {
        "properties": {
          "accepted": {
            "example": true,
            "type": "boolean"
          }
        },
        "required": ["accepted"],
        "type": "object"
      },
...
      "OffboardingCountsResponseDto": {
        "properties": {
          "openTasks": {
            "type": "number"
          },
          "records": {
            "type": "number"
          }
        },
        "required": ["records", "openTasks"],
        "type": "object"
      },
      "OffboardingPreviewResponseDto": {
        "properties": {
          "openTasks": {
            "type": "number"
          },
          "recipients": {
            "items": {
              "$ref": "#/components/schemas/TenantMemberResponseDto"
            },
            "type": "array"
          },
          "records": {
            "type": "number"
          }
        },
        "required": ["records", "openTasks", "recipients"],
        "type": "object"
      },
...
      "TenantInvitationResponseDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "role": {
            "enum": ["TENANT_ADMIN", "EMPLOYEE"],
            "type": "string"
          },
          "status": {
            "enum": ["PENDING", "ACCEPTED", "DECLINED", "REVOKED", "EXPIRED"],
            "type": "string"
          },
          "targetPhone": {
            "example": "+8613800138000",
            "type": "string"
          },
          "targetUserId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "targetPhone",
          "role",
          "status",
          "expiresAt",
          "createdAt"
        ],
        "type": "object"
      },
      "TenantMemberPageResponseDto": {
        "properties": {
          "activeAdminCount": {
            "minimum": 0,
            "type": "number"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TenantMemberResponseDto"
            },
            "type": "array"
          },
          "limit": {
            "maximum": 100,
            "minimum": 1,
            "type": "number"
          },
          "page": {
            "minimum": 1,
            "type": "number"
          },
          "total": {
            "minimum": 0,
            "type": "number"
          }
        },
        "required": ["items", "page", "limit", "total", "activeAdminCount"],
        "type": "object"
      },
      "TenantMemberResponseDto": {
        "properties": {
          "displayName": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "phone": {
            "example": "+8613800138000",
            "type": "string"
          },
          "role": {
            "enum": ["TENANT_ADMIN", "EMPLOYEE"],
            "type": "string"
          },
          "status": {
            "enum": ["ACTIVE", "DISABLED"],
            "type": "string"
          },
          "tenantId": {
            "format": "uuid",
            "type": "string"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": ["id", "userId", "tenantId", "role", "status"],
        "type": "object"
      },
...
      "WorkspaceSummaryResponseDto": {
        "properties": {
          "memberId": {
            "format": "uuid",
            "type": "string"
          },
          "memberStatus": {
            "enum": ["ACTIVE", "DISABLED"],
            "type": "string"
          },
          "role": {
            "enum": ["TENANT_ADMIN", "EMPLOYEE"],
            "type": "string"
          },
          "tenantCode": {
            "type": "string"
          },
          "tenantId": {
            "format": "uuid",
            "type": "string"
          },
          "tenantName": {
            "type": "string"
          },
          "tenantStatus": {
            "enum": ["DRAFT", "ACTIVE", "SUSPENDED", "CLOSED"],
            "type": "string"
          }
        },
        "required": [
          "tenantId",
          "tenantCode",
          "tenantName",
          "tenantStatus",
          "memberId",
          "memberStatus",
          "role"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "crm_session": {
        "in": "cookie",
        "name": "crm_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "多租户 CRM 平台 REST API",
    "title": "Multi-tenant CRM API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/me/workspaces": {
      "get": {
        "operationId": "MembershipsController_workspaces",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceSummaryResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["memberships"]
      }
    },
...
    "/api/v1/workspaces/{tenantCode}": {
      "get": {
        "operationId": "MembershipsController_workspace",
        "parameters": [
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSummaryResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["memberships"]
      }
    },
...
    "/api/v1/workspaces/{tenantCode}/invitations": {
      "get": {
        "operationId": "MembershipsController_invitations",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationPageResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["memberships"]
      },
      "post": {
        "operationId": "MembershipsController_invite",
        "parameters": [
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedInvitationResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["memberships"]
      }
    },
    "/api/v1/workspaces/{tenantCode}/invitations/{id}/resend": {
      "post": {
        "operationId": "MembershipsController_resend",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResentResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["memberships"]
      }
    },
    "/api/v1/workspaces/{tenantCode}/invitations/{id}/revoke": {
      "post": {
        "operationId": "MembershipsController_revoke",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipActionResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["memberships"]
      }
    },
    "/api/v1/workspaces/{tenantCode}/members": {
      "get": {
        "operationId": "MembershipsController_members",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantMemberPageResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["memberships"]
      }
    },
    "/api/v1/workspaces/{tenantCode}/members/admin-handoff": {
      "post": {
        "operationId": "MembershipsController_handoff",
        "parameters": [
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberRecipientDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipActionResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["memberships"]
      }
    },
    "/api/v1/workspaces/{tenantCode}/members/{memberId}": {
      "patch": {
        "operationId": "MembershipsController_changeMember",
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeMemberStatusDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantMemberResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["memberships"]
      }
    },
    "/api/v1/workspaces/{tenantCode}/members/{memberId}/object-access": {
      "get": {
        "operationId": "MembershipsController_memberObjectAccess",
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemberObjectAccessResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["memberships"]
      }
    },
    "/api/v1/workspaces/{tenantCode}/members/{memberId}/object-access/{objectId}": {
      "put": {
        "operationId": "MembershipsController_setMemberObjectAccess",
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "objectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberObjectAccessDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberObjectAccessResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["memberships"]
      }
    },
    "/api/v1/workspaces/{tenantCode}/members/{memberId}/offboarding": {
      "get": {
        "operationId": "MembershipsController_previewOffboarding",
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OffboardingPreviewResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["memberships"]
      },
      "post": {
        "operationId": "MembershipsController_offboard",
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberRecipientDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OffboardingCountsResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["memberships"]
      }
    },
    "/api/v1/workspaces/{tenantCode}/members/{memberId}/role": {
      "patch": {
        "operationId": "MembershipsController_changeRole",
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeMemberRoleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantMemberResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["memberships"]
      }
    },
...
  },
  "servers": [],
...
}